HYDRAN-2069: switch decimal API fields to BigDecimal + small annotation cleanup
Money and other decimal fields in the public API (Invoice, CustomerInvoice,
InvoiceDetail) were stored as float, which is the wrong type for currency:
precision drift, no exact decimal, and primitives can't be null so missing
upstream values silently became 0 (indistinguishable from a real "0 SEK"
amount). DWR already gives us BigDecimal (or Double for percentages and
quantities), so the mapper was actively narrowing the data and substituting
ZERO for nulls.

Change all amount, vat, vatRate, quantity, unitPrice and rounding fields to
BigDecimal. The mapper becomes pass-through for upstream BigDecimal values,
and converts upstream Doubles (vatRate, quantity) via toBigDecimal. Nulls
flow end-to-end now — @JsonInclude(NON_NULL) omits the field when missing,
so consumers can distinguish "missing" from "zero". The synthesized
amountVatIncluded for invoicecache uses a small nullableAdd helper to
preserve null-when-both-null while still adding when at least one is set.

Also in this commit:
- InvoicesResource @ApiResponse(200) annotations now use
  useReturnTypeSchema = true so the schema can't drift from the actual
  return type.
- Drop the no-op @ValidOrganizationNumber(nullable = true) on
  organizationNumbers/organizationIds list elements; query-param binding
  never produces null entries, so the flag was misleading.
- Refresh src/main/resources/openapi.yaml from the regenerated spec
  (number type changes) and prune now-omitted vatEligibleAmount/rounding
  zeros from the GetInvoicesPublicAdministration IT fixtures.

Acceptable as a breaking change because the service is on the new 9.2
major and no external consumers depend on this branch's shape yet.

+++ src/test/java/se/sundsvall/invoices/service/InvoicesServiceTest.java
@@ -200,8 +200,8 @@ class InvoicesServiceTest {
 		final var organizationNumber = "5523456789";
 		final var invoiceNumber = "111222";
 		final var expectedInvoiceDetail = new InvoiceDetail();
-		expectedInvoiceDetail.setAmount(10.45f);
-		expectedInvoiceDetail.setQuantity(2);
+		expectedInvoiceDetail.setAmount(BigDecimal.valueOf(10.45));
+		expectedInvoiceDetail.setQuantity(BigDecimal.valueOf(2d));
 
 		when(dataWarehouseReaderClientMock.getInvoiceDetails(municipalityId, organizationNumber, Long.parseLong(invoiceNumber))).thenReturn(List.of(createDataWarehouseReaderInvoiceDetail(invoiceNumber)));
